+++ c3IDE/Windows/TypeWindow.xaml.cs
@@ -35,6 +35,7 @@ namespace c3IDE.Windows
         private CompletionWindow completionWindow;
         private FoldingManager edittimeFoldingManager, runtimeFoldingManager;
         private BraceFoldingStrategy folding;
+        private SearchPanel edittimePanel, runtimePanel;
 
         /// <summary>
         /// constructor for type window
@@ -60,8 +61,8 @@ namespace c3IDE.Windows
             folding.UpdateFoldings(runtimeFoldingManager, RunTimeTypeTextEditor.Document);
 
             //setip ctrl-f to single page code find
-            SearchPanel.Install(EditTimeTypeTextEditor);
-            SearchPanel.Install(EditTimeTypeTextEditor);
+            edittimePanel = SearchPanel.Install(EditTimeTypeTextEditor);
+            runtimePanel = SearchPanel.Install(EditTimeTypeTextEditor);
         }
 
         /// <summary>
@@ -71,6 +72,7 @@ namespace c3IDE.Windows
         {
             ThemeManager.SetupTextEditor(EditTimeTypeTextEditor, Syntax.Javascript);
             ThemeManager.SetupTextEditor(RunTimeTypeTextEditor, Syntax.Javascript);
+            ThemeManager.SetupSearchPanel(edittimePanel, runtimePanel);
 
             if (AddonManager.CurrentAddon != null)
             {
@@ -320,6 +322,47 @@ namespace c3IDE.Windows
             foreach (var fold in runtimeFoldingManager.AllFoldings)
             {
                 fold.IsFolded = false;
+            }
+        }
+
+        private void FindGlobal_Click(object sender, RoutedEventArgs e)
+        {
+            //AppData.Insatnce.GlobalSave(false);
+            Searcher.Insatnce.UpdateFileIndex("edittime_type.js", EditTimeTypeTextEditor.Text, ApplicationWindows.InstanceWindow);
+            Searcher.Insatnce.UpdateFileIndex("runtime_type.js", RunTimeTypeTextEditor.Text, ApplicationWindows.InstanceWindow);
+
+            MenuItem mnu = sender as MenuItem;
+            TextEditor editor = null;
+
+            if (mnu != null)
+            {
+                editor = ((ContextMenu)mnu.Parent).PlacementTarget as TextEditor;
+                var text = editor.SelectedText;
+                Searcher.Insatnce.GlobalFind(text, this);
+            }
+        }
+
+        private void CommentSelection(object sender, RoutedEventArgs e)
+        {
+            MenuItem mnu = sender as MenuItem;
+            TextEditor editor = null;
+
+            if (mnu != null)
+            {
+                editor = ((ContextMenu)mnu.Parent).PlacementTarget as TextEditor;
+                editor.CommentSelectedLines();
+            }
+        }
+
+        private void UncommentSelection(object sender, RoutedEventArgs e)
+        {
+            MenuItem mnu = sender as MenuItem;
+            TextEditor editor = null;
+
+            if (mnu != null)
+            {
+                editor = ((ContextMenu)mnu.Parent).PlacementTarget as TextEditor;
+                editor.UncommentSelectedLines();
             }
         }
     }